requirements.txt
fastapi
uvicorn
streamlit
sentence-transformers
scikit-learn
pdfplumber
python-docx
pandas
numpy
python-multipart